DDF-3599 Moved generic matchers to test common library
