+++ libs/test-common/src/main/java/org/codice/ddf/test/common/matchers/CastingMatchers.java
@@ -11,7 +11,7 @@
  * License is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package org.codice.ddf.test.matchers;
+package org.codice.ddf.test.common.matchers;
 
 import org.apache.commons.lang.ClassUtils;
 import org.hamcrest.Description;
