+++ platform/migration/platform-migration/src/test/java/org/codice/ddf/configuration/migration/ImportMigrationContextImplTest.java
@@ -28,8 +28,8 @@ import org.codice.ddf.migration.Migratable;
 import org.codice.ddf.migration.MigrationEntry;
 import org.codice.ddf.migration.MigrationException;
 import org.codice.ddf.migration.MigrationOperation;
-import org.codice.ddf.test.matchers.CastingMatchers;
-import org.codice.ddf.test.matchers.MappingMatchers;
+import org.codice.ddf.test.common.matchers.CastingMatchers;
+import org.codice.ddf.test.common.matchers.MappingMatchers;
 import org.hamcrest.Matchers;
 import org.junit.Assert;
 import org.junit.Before;
