+++ platform/migration/platform-migration/src/test/java/org/codice/ddf/configuration/migration/ImportMigrationPropertyReferencedEntryImplTest.java
@@ -27,7 +27,7 @@ import org.codice.ddf.migration.ImportMigrationEntry;
 import org.codice.ddf.migration.MigrationException;
 import org.codice.ddf.migration.MigrationOperation;
 import org.codice.ddf.migration.MigrationReport;
-import org.codice.ddf.test.matchers.ThrowableMatchers;
+import org.codice.ddf.test.common.matchers.ThrowableMatchers;
 import org.codice.ddf.util.function.BiThrowingConsumer;
 import org.hamcrest.Matchers;
 import org.junit.Assert;
