+++ platform/migration/platform-migration/src/test/java/org/codice/ddf/configuration/migration/MigrationReportImplTest.java
@@ -25,8 +25,8 @@ import org.codice.ddf.migration.MigrationMessage;
 import org.codice.ddf.migration.MigrationOperation;
 import org.codice.ddf.migration.MigrationReport;
 import org.codice.ddf.migration.MigrationWarning;
-import org.codice.ddf.test.matchers.MappingMatchers;
-import org.codice.ddf.test.matchers.ThrowableMatchers;
+import org.codice.ddf.test.common.matchers.MappingMatchers;
+import org.codice.ddf.test.common.matchers.ThrowableMatchers;
 import org.codice.ddf.util.function.ThrowingRunnable;
 import org.hamcrest.Matchers;
 import org.junit.Assert;
